add brokerage types to fsharp project
