+++ src/core.fs/core.fs.fsproj
@@ -13,6 +13,8 @@
         <Compile Include="Shared\Adapters\Storage\IPortfolioStorage.fs"/>
         <Compile Include="Shared\Adapters\Subscriptions\Types.fs" />
         <Compile Include="Shared\Adapters\Logging.fs" />
+        <Compile Include="Shared\Adapters\Brokerage\Types.fs" />
+        <Compile Include="Services\Trading.fs" />
         <Compile Include="Options\Input.fs"/>
         <Compile Include="Options\Views.fs"/>
         <Compile Include="Options\Handler.fs"/>
