+++ src/core/Shared/Ticker.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace core.Shared
 {
-    public struct Ticker
+    public struct Ticker : IComparable
     {
         private readonly string _ticker;
 
@@ -19,5 +19,14 @@ namespace core.Shared
         public static implicit operator Ticker(string t) => new Ticker(t);
 
         public string Value => _ticker;
+        public int CompareTo(object obj)
+        {
+            if (obj is Ticker t)
+            {
+                return string.CompareOrdinal(_ticker, t._ticker);
+            }
+
+            return -1;
+        }
     }
 }
