+++ src/web/BackgroundServices/EmailNotificationService.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using core.fs.Alerts;
+using core.fs.Shared.Adapters.Brokerage;
 using core.fs.Shared.Adapters.Storage;
-using core.Shared.Adapters.Brokerage;
 using core.Shared.Adapters.Emails;
 using Microsoft.Extensions.Logging;
