+++ tests/coretests/Options/DashboardTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using core.Account;
 using core.fs.Options;
+using core.fs.Shared.Adapters.Brokerage;
 using core.fs.Shared.Adapters.Storage;
 using core.fs.Shared.Domain.Accounts;
 using core.Shared;
@@ -39,7 +40,7 @@ namespace coretests.Options
                 );
 
             var brokerage = new Mock<IBrokerage>();
-            brokerage.Setup(x => x.GetQuotes(It.IsAny<UserState>(), It.IsAny<List<string>>()))
+            brokerage.Setup(x => x.GetQuotes(It.IsAny<UserState>(), It.IsAny<List<Ticker>>()))
                 .Returns(Task.FromResult(
                     new ServiceResponse<Dictionary<string, StockQuote>>(
                         new Dictionary<string, StockQuote>()
