+++ tests/infrastructuretests/timezonesupporttests/MarketHoursTests.cs
@@ -1,4 +1,4 @@
-using core.Shared.Adapters.Brokerage;
+using core.fs.Shared.Adapters.Brokerage;
 
 namespace timezonesupporttests;
 public class MarketHoursTests
